requirements.txt
pandas
requests
openpyxl
plotly
streamlit
urllib3
Pillow
altair==4.2.2
./imghdr_shim